levelpage.js dosyası parcalara bolundu fonksıyonelluk artılırdı artık jerseyı import edıyor levelpage css dosyası herseyi import ediyor fonksiyonellık arttıtldı gpt sagolsun
