+++ frontend/css/components/lessonCard.css
@@ -1,0 +1,89 @@
+.lesson-card {
+    background: rgba(24, 28, 40, 0.92);
+    border-radius: 1.2rem;
+    box-shadow: 0 4px 18px 0 rgba(16, 24, 40, 0.18);
+    padding: 1.3rem 1.1rem 1rem 1.1rem;
+    min-width: 240px;
+    max-width: 320px;
+    width: 100%;
+    color: #f1f5f9;
+    margin: 0;
+    border: 1.5px solid rgba(37, 99, 235, 0.13);
+    transition: transform 0.18s, box-shadow 0.18s, background 0.18s;
+    cursor: pointer;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    animation: fadeIn 0.5s cubic-bezier(.4, 0, .2, 1);
+    backdrop-filter: blur(1.5px);
+}
+
+.lesson-card:hover {
+    transform: translateY(-6px) scale(1.035);
+    box-shadow: 0 12px 36px 0 rgba(37, 99, 235, 0.22);
+    background: rgba(37, 99, 235, 0.18);
+}
+
+.lesson-number {
+    font-size: 1.02rem;
+    color: #60a5fa;
+    font-weight: 600;
+    margin-bottom: 0.2rem;
+    letter-spacing: 0.01em;
+}
+
+.lesson-title {
+    font-size: 1.18rem;
+    font-weight: 700;
+    color: #fbbf24;
+    margin-bottom: 0.5rem;
+    letter-spacing: 0.01em;
+}
+
+.lesson-description {
+    font-size: 1.04rem;
+    color: #cbd5e1;
+    margin-bottom: 0.2rem;
+    line-height: 1.5;
+    word-break: break-word;
+}
+
+.lesson-footer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    margin-top: 0.7rem;
+    gap: 1rem;
+}
+
+.lesson-stars {
+    display: flex;
+    gap: 0.15em;
+}
+
+.star {
+    color: #fbbf24;
+    font-size: 1.1em;
+    filter: drop-shadow(0 0 2px #fff2);
+}
+
+.lesson-wpm {
+    font-size: 1em;
+    color: #60a5fa;
+    font-weight: 600;
+    min-width: 2.5em;
+    text-align: right;
+}
+
+@keyframes fadeIn {
+    from {
+        opacity: 0;
+        transform: translateY(24px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
