+++ frontend/css/components/userPanel.css
@@ -1,0 +1,78 @@
+.user-stats-container {
+    max-width: 600px;
+    margin: 2.5rem auto 1.5rem auto;
+    border-radius: 1.2rem;
+    box-shadow: 0 8px 32px 0 rgba(37, 99, 235, 0.13);
+    background: rgba(255, 255, 255, 0.18);
+    padding: 0.5rem 0 1.2rem 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0;
+}
+
+.user-info-panel {
+    max-width: 400px;
+    margin: 2rem auto 1.5rem auto;
+    padding: 1rem 2rem;
+    border-radius: 1rem;
+    background: rgba(52, 37, 37, 0.25);
+    backdrop-filter: blur(2px);
+    box-shadow: 0 8px 32px 0 rgba(255, 255, 255, 0.37);
+    text-align: center;
+    font-size: 2.15rem;
+    font-weight: 500;
+    transition: background 0.3s;
+    margin-bottom: 2rem;
+}
+
+.user-info-panel span {
+    font-weight: bold;
+    color: #10b981;
+}
+
+.progress-stats {
+    padding: 1.5rem;
+    margin-bottom: 2rem;
+}
+
+.stats-grid {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 1rem;
+    text-align: center;
+}
+
+.stat-item {
+    padding: 0.5rem;
+}
+
+.best-wpm-text {
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: #3b82f6;
+    margin-bottom: 0.25rem;
+}
+
+.progress-text {
+    font-size: 1.5rem;
+    font-weight: bold;
+    margin-bottom: 0.5rem;
+    color: #3b82f6;
+}
+
+.progress-bar {
+    width: 100%;
+    height: 0.5rem;
+    background: rgba(200, 200, 200, 0.3);
+    border-radius: 0.25rem;
+    overflow: hidden;
+    margin-top: 0.5rem;
+}
+
+.progress-fill {
+    height: 100%;
+    background: #10b981;
+    border-radius: 0.25rem;
+    transition: width 0.3s ease;
+}
