+++ frontend/css/levelpage.css
@@ -1,41 +1,35 @@
-.user-stats-container {
-    max-width: 600px;
-    margin: 2.5rem auto 1.5rem auto;
-    border-radius: 1.2rem;
-    box-shadow: 0 8px 32px 0 rgba(37, 99, 235, 0.13);
-    background: rgba(255, 255, 255, 0.18);
-    padding: 0.5rem 0 1.2rem 0;
+@import url("./components/layout.css");
+@import url("./components/userPanel.css");
+@import url("./components/lessonCard.css");
+@import url("./components/adminPanel.css");
+
+/* Ders Kartı Alt Bilgisi ve Yıldızlar */
+.lesson-footer {
     display: flex;
-    flex-direction: column;
+    justify-content: space-between;
     align-items: center;
-    gap: 0;
-}
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    min-height: 100vh;
-    overflow-x: hidden;
-    color: #333;
-    transition: color 0.3s ease;
-}
-
-.page-background {
-    position: fixed;
-    top: 0;
-    left: 0;
     width: 100%;
-    height: 100%;
-    z-index: -1;
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    transition: opacity 1s ease;
+    margin-top: 0.7rem;
+    gap: 1rem;
+}
+
+.lesson-stars {
+    display: flex;
+    gap: 0.15em;
+}
+
+.star {
+    color: #fbbf24;
+    font-size: 1.1em;
+    filter: drop-shadow(0 0 2px #fff2);
+}
+
+.lesson-wpm {
+    font-size: 1em;
+    color: #60a5fa;
+    font-weight: 600;
+    min-width: 2.5em;
+    text-align: right;
 }
 
 .page-background.light {
@@ -46,33 +40,11 @@ body {
     background-image: url("../images/backgrnd2.jpg");
 }
 
-
-
-.glass-card {
-    backdrop-filter: blur(3px);
-    background: rgba(255, 255, 255, 0.25);
-    border-radius: 18px;
-    box-shadow: 0 8px 32px 0 rgba(255, 255, 255, 0.37);
-    border: 1px solid rgba(255, 255, 255, 0.18);
-    transition: all 0.3s ease;
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
 }
-
-.glass-header {
-    backdrop-filter: blur(3px);
-    background: rgba(255, 255, 255, 0.1);
-    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-    position: sticky;
-    top: 0;
-    z-index: 50;
-}
-
-.glass-footer {
-    backdrop-filter: blur(3px);
-    background: rgba(255, 255, 255, 0.1);
-    border-top: 1px solid rgba(255, 255, 255, 0.1);
-    margin-top: 4rem;
-}
-
 
 .navbar {
     display: flex;
@@ -143,148 +115,6 @@ body {
 }
 
 
-
-
-
-.main-content {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 2rem 1rem;
-}
-
-.progress-stats {
-    padding: 1.5rem;
-    margin-bottom: 2rem;
-}
-
-.stats-grid {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 1rem;
-    text-align: center;
-}
-
-.stat-item {
-    padding: 0.5rem;
-}
-
-.best-wpm-text {
-    font-size: 1.5rem;
-    font-weight: bold;
-    color: #3b82f6;
-    margin-bottom: 0.25rem;
-}
-
-.progress-text {
-    font-size: 1.5rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-    color: #3b82f6;
-
-}
-
-.progress-bar {
-    width: 100%;
-    height: 0.5rem;
-    background: rgba(200, 200, 200, 0.3);
-    border-radius: 0.25rem;
-    overflow: hidden;
-    margin-top: 0.5rem;
-}
-
-.progress-fill {
-    height: 100%;
-    background: #10b981;
-    border-radius: 0.25rem;
-    transition: width 0.3s ease;
-}
-
-.lesson-sections {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 2.5rem;
-    justify-content: center;
-    align-items: flex-start;
-    margin-top: 0.5rem;
-    margin-bottom: 3rem;
-    width: 100%;
-}
-
-.lesson-section {
-    margin-bottom: 2rem;
-}
-
-.section-title {
-    font-size: 1.5rem;
-    font-weight: 700;
-    color: #fff;
-    margin: 2.5rem 0 0.5rem 2.5rem;
-    /* Bunu azaltabilirsin */
-    letter-spacing: 0.01em;
-}
-
-.section-icon {
-    font-size: 1.25rem;
-}
-
-.lesson-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-    gap: 1.5rem;
-    width: 100%;
-    justify-items: center;
-}
-
-.lesson-card {
-    background: rgba(24, 28, 40, 0.92);
-    border-radius: 1.2rem;
-    box-shadow: 0 4px 18px 0 rgba(16, 24, 40, 0.18);
-    padding: 1.3rem 1.1rem 1rem 1.1rem;
-    min-width: 240px;
-    max-width: 320px;
-    width: 100%;
-    color: #f1f5f9;
-    margin: 0;
-    border: 1.5px solid rgba(37, 99, 235, 0.13);
-    transition: transform 0.18s, box-shadow 0.18s, background 0.18s;
-    cursor: pointer;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    animation: fadeIn 0.5s cubic-bezier(.4, 0, .2, 1);
-    backdrop-filter: blur(1.5px);
-}
-
-.lesson-card:hover {
-    transform: translateY(-6px) scale(1.035);
-    box-shadow: 0 12px 36px 0 rgba(37, 99, 235, 0.22);
-    background: rgba(37, 99, 235, 0.18);
-}
-
-.lesson-number {
-    font-size: 1.02rem;
-    color: #60a5fa;
-    font-weight: 600;
-    margin-bottom: 0.2rem;
-    letter-spacing: 0.01em;
-}
-
-.lesson-title {
-    font-size: 1.18rem;
-    font-weight: 700;
-    color: #fbbf24;
-    margin-bottom: 0.5rem;
-    letter-spacing: 0.01em;
-}
-
-.lesson-description {
-    font-size: 1.04rem;
-    color: #cbd5e1;
-    margin-bottom: 0.2rem;
-    line-height: 1.5;
-    word-break: break-word;
-}
-
 @keyframes fadeIn {
     from {
         opacity: 0;
@@ -315,138 +145,4 @@ body {
         max-width: 98vw;
         padding: 1.1rem 0.7rem;
     }
-}
-
-.user-info-panel {
-    max-width: 400px;
-    margin: 2rem auto 1.5rem auto;
-    padding: 1rem 2rem;
-    border-radius: 1rem;
-    background: rgba(52, 37, 37, 0.25);
-    backdrop-filter: blur(2px);
-    box-shadow: 0 8px 32px 0 rgba(255, 255, 255, 0.37);
-    text-align: center;
-    font-size: 2.15rem;
-    font-weight: 500;
-    transition: background 0.3s;
-    margin-bottom: 2rem;
-}
-
-.user-info-panel span {
-    font-weight: bold;
-    color: #10b981;
-}
-
-#adminDropdown {
-    display: none;
-    position: absolute;
-    top: 110%;
-    left: 0;
-    background: rgba(255, 255, 255, 0.98);
-    border-radius: 8px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-    min-width: 150px;
-    z-index: 100;
-    padding: 0.25rem 0;
-    display: none;
-    position: absolute;
-    top: 110%;
-    left: 0;
-    background: rgba(255, 255, 255, 0.95);
-    border-radius: 8px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-    min-width: 150px;
-    z-index: 100;
-}
-
-#adminDropdown li {
-    list-style: none;
-}
-
-#adminDropdown button {
-    width: 100%;
-    background: none;
-    border: none;
-    padding: 10px 16px;
-    text-align: left;
-    cursor: pointer;
-    font-size: 1em;
-    color: #2563eb;
-    border-radius: 0;
-    transition: background 0.2s;
-}
-
-#adminDropdown button:hover {
-    background: #f3f4f6;
-}
-
-#addLevelModal {
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background: rgba(0, 0, 0, 0.25);
-    z-index: 9999;
-    align-items: center;
-    justify-content: center;
-
-
-}
-
-#addLevelModal.show {
-    display: flex;
-}
-
-#addLevelForm {
-    background: #fff;
-    border-radius: 12px;
-    padding: 2rem 2.5rem;
-    min-width: 320px;
-    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.18);
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-
-
-}
-
-#addLevelForm input,
-#addLevelForm textarea {
-    padding: 0.5rem;
-    border-radius: 6px;
-    border: 1px solid #ccc;
-    font-size: 1em;
-}
-
-#addLevelForm button[type="submit"] {
-    background: #10b981;
-    color: #fff;
-    border: none;
-    border-radius: 6px;
-    padding: 0.5rem 1.2rem;
-    cursor: pointer;
-}
-
-#addLevelForm button[type="button"] {
-    background: #eee;
-    border: none;
-    border-radius: 6px;
-    padding: 0.5rem 1.2rem;
-    cursor: pointer;
-}
-
-#addLevelMsg {
-    color: #e11d48;
-    font-size: 0.95em;
-    min-height: 1.2em;
-}
-
-#addLevelMsg {
-    width: 100%;
-    background: none;
-    border: none;
-    padding: 10px 16px;
-    text-align: left;
-    cursor: pointer;
-    font-size: 1em;
 }
