+++ frontend/html/levelpage.html
@@ -110,7 +110,7 @@
 
     </main>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script src="../js/levelpage.js"></script>
+    <script type="module" src="../js/levelpage.js"></script>
     <script src="../js/adminControlpanel.js"></script>
 
 </body>
