+++ frontend/js/levelpage.js
@@ -1,242 +1,49 @@
-// Theme Management
-class ThemeManager {
-  constructor() {
-    this.currentTheme = localStorage.getItem("theme") || "light";
-    this.init();
-  }
+import { ThemeManager } from "../utils/themeManager.js";
+import { AnimationManager } from "../utils/animationManager.js";
+import { fetchUserData, setProgress } from "../components/userPanel.js";
+import { checkLogin } from "../utils/auth.js";
+import { createLessonCard } from "../components/lessonCard.js";
 
-  init() {
-    this.applyTheme(this.currentTheme);
-    this.setupEventListeners();
-  }
-
-  setupEventListeners() {
-    const themeToggle = document.getElementById("themeToggle");
-    if (themeToggle) {
-      themeToggle.addEventListener("click", () => this.toggleTheme());
-    }
-  }
-
-  toggleTheme() {
-    this.currentTheme = this.currentTheme === "light" ? "dark" : "light";
-    this.applyTheme(this.currentTheme);
-    localStorage.setItem("theme", this.currentTheme);
-  }
-
-  applyTheme(theme) {
-    const body = document.body;
-    const background = document.querySelector(".page-background");
-    const themeIcon = document.getElementById("themeIcon");
-
-    // Remove existing theme classes
-    body.classList.remove("light", "dark");
-    if (background) {
-      background.classList.remove("light", "dark");
-    }
-
-    // Apply new theme
-    body.classList.add(theme);
-    if (background) {
-      background.classList.add(theme);
-    }
-
-    // Update theme icon
-    if (themeIcon) {
-      themeIcon.textContent = theme === "dark" ? "☀️" : "🌙";
-    }
-  }
-}
-
-// Animation and Effects
-class AnimationManager {
-  static addFadeIn(element) {
-    element.classList.add("fade-in");
-    setTimeout(() => element.classList.remove("fade-in"), 500);
-  }
-
-  static addSlideUp(element) {
-    element.classList.add("slide-up");
-    setTimeout(() => element.classList.remove("slide-up"), 300);
-  }
-
-  static smoothScrollTo(element) {
-    element.scrollIntoView({
-      behavior: "smooth",
-      block: "center",
-    });
-  }
-}
-
+// Uygulama başlatıcı
 class TypingTutorApp {
   constructor() {
     this.themeManager = new ThemeManager();
-
     this.init();
   }
-
   init() {
-
-    // Add fade-in animation to main content
     const mainContent = document.querySelector(".main-content");
     if (mainContent) {
       AnimationManager.addFadeIn(mainContent);
     }
-
   }
-
 }
-// Initialize app when DOM is loaded
-document.addEventListener("DOMContentLoaded", () => {
+
+// Sayfa yüklendiğinde çalışacak
+document.addEventListener("DOMContentLoaded", async () => {
   window.typingApp = new TypingTutorApp();
+  await checkLogin();// giriş kontrolcüsü
+  await fetchUserData();/// kullanıcı verilerini al
+  loadLessons(); // dersleri yükle
 });
-
-
 
 document.querySelector(".login-btn").onclick = function () {
-  window.location.href = "login.html";
+  window.location.href = "login.html"; // Giriş sayfasına yönlendir
 };
-
-
-// /me endpoint'inden kullanıcı verilerini çek
-async function fetchUserData() {
-  const progressElement = document.querySelector(".progress-text");
-  const bestWPMElement = document.querySelector(".best-wpm-text");
-  const userNameElement = document.getElementById("userNameDisplay");
-
-  try {
-    // Kullanıcı verilerini çek
-    const response = await axios.get(`http://localhost:4000/database/me`, {
-      withCredentials: true,
-    });
-
-    const userData = response.data;
-    console.log("Kullanıcı Verileri:", userData);
-    if (userData.data.role === "admin") {
-      document.getElementById("adminPanelBtn").style.display = "block";
-    }
-
-    // Progress bar güncelle
-    setProgress(userData.data.completionStats.percentage);
-
-    if (progressElement) {
-      progressElement.textContent = `${userData.data.completionStats.percentage}%`;
-    }
-    if (bestWPMElement) {
-      bestWPMElement.textContent = `${userData.data.topWPM} WPM`;
-    }
-    if (userNameElement) {
-      userNameElement.textContent = userData.data.name || "Misafir";
-    }
-
-  } catch (error) {
-    const errMsg = error.response.msg;
-    console.error("Hata:", errMsg);// butası calısmıyor
-
-    if (progressElement) progressElement.textContent = "Hata! " + errMsg;
-    if (bestWPMElement) bestWPMElement.textContent = "Hata! " + errMsg;
-    if (userNameElement) userNameElement.textContent = "Hata! " + errMsg;
-  }
-}
-
-const checkLogin = async () => {
-  try {
-    const response = await axios.get(`http://localhost:4000/auth/check`, {
-      withCredentials: true
-    });
-
-    const loginBtn = document.getElementById("loginButton");
-    const logoutBtn = document.getElementById("logoutButton");
-
-    if (response.data.loggedIn) {
-      console.log("Kullanıcı giriş yapmış");
-      if (loginBtn) loginBtn.style.display = "none";
-      if (logoutBtn) {
-        logoutBtn.style.display = "block";
-        logoutBtn.onclick = async function () {
-          await axios.post(`http://localhost:4000/auth/logout`, {}, {
-            withCredentials: true
-          }).then(() => {
-            console.log("Çıkış yapıldı");
-            window.location.href = "login.html";
-            checkLogin();
-          }).catch(error => {
-            console.error("Çıkış yaparken hata:", error);
-          });
-        };
-      }
-    } else {
-      console.log("Kullanıcı giriş yapmamış");
-      if (loginBtn) loginBtn.style.display = "block";
-      if (logoutBtn) logoutBtn.style.display = "none";
-    }
-  } catch (err) {
-    console.error("Check isteğinde hata:", err);
-  }
-};
-
-
-
-// Sayfa yüklendiğinde giriş durumunu kontrol et
-document.addEventListener("DOMContentLoaded", async () => {
-  await checkLogin();
-  await fetchUserData();
-
-  // Giriş Yap butonu: login.html'ye yönlendir
-  const loginButton = document.getElementById("loginButton");
-  if (loginButton) {
-    loginButton.onclick = () => {
-      window.location.href = "login.html";
-    };
-  }
-
-  // Çıkış Yap butonu: LocalStorage'ı temizle ve butonları güncelle
-  /* const logoutButton = document.getElementById("logoutButton");
-   if (logoutButton) {
-     logoutButton.onclick = () => {
-       localStorage.removeItem("isLoggedIn");
-       localStorage.removeItem("token");
-       checkLoginStatus();
-     };
-   }*/
-});
-
-
-//genel ilerleme barı
-function setProgress(percent) {
-  percent = Math.max(0, Math.min(100, percent)); // 0-100 arası sınırla
-  document.getElementById('progressText').textContent = percent + '%';
-  document.getElementById('progressFill').style.width = percent + '%';
-}
 
 // Dersleri yükle
 async function loadLessons() {
   try {
-    const res = await axios.get("http://localhost:4000/database/all-levels");
-    // Eğer res.data bir nesne ve içinde dizi varsa(gpt):
+    const res = await axios.get("http://localhost:4000/database/all-levels");// Tüm dersleri al
+    // Gelen verileri kontrol et(burwsı gptden)
     const lessons = Array.isArray(res.data) ? res.data : res.data.data || [];
-    const container = document.getElementById("lessons-container");
-    container.innerHTML = "";
-
-    console.log("Gelen lessons verisi:", lessons);
-
+    const container = document.getElementById("lessons-container");// Dersler konteynerini al
+    container.innerHTML = "";// Önceki dersleri temizle
     lessons.forEach((lesson, idx) => {
-      const isAvailable = lesson.isAvailable || idx < 2; // örnek: ilk 2 ders açık, diğerleri kilitli
-      const card = document.createElement("div");
-      card.className = `lesson-card ${isAvailable ? 'lesson-card-available' : 'lesson-card-locked'}`;
-      card.innerHTML = `
-        <div class="lesson-content">
-          <div class="lesson-number">${idx + 1}</div>
-          <div class="lesson-icon">${isAvailable ? '▶️' : '🔒'}</div>
-          <h3 class="lesson-title">${lesson.name || ''}</h3>
-          <p class="lesson-description">${lesson.description || ''}</p>
-          <div class="lesson-footer">
-            <div class="lesson-wpm">${isAvailable ? `${lesson.__v}` : '🔒'}</div>
-          </div>
-        </div>
-      `;
-      if (isAvailable) {
-        card.onclick = () => startLesson(idx + 1);
-      }
+      // Her bir ders için kart oluştur
+      const isAvailable = lesson.isAvailable || idx < 2;
+      // Kart oluşturma fonksiyonunu çağır
+      const card = createLessonCard(lesson, idx, isAvailable, startLesson);
+      // Kartı konteynere ekle
       container.appendChild(card);
     });
   } catch (err) {
@@ -244,4 +51,7 @@ async function loadLessons() {
   }
 }
 
-document.addEventListener("DOMContentLoaded", loadLessons);
+function startLesson(idx) {
+  //eklenecekk
+  alert(`Ders ${idx} başlatılacak!`);
+}
